requirements.txt
nltk==3.8.1
pandas==1.5.3
pdfminer.six==20221105
seaborn==0.12.2
wordcloud==1.8.2.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
